+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 0.64 (unreleased)
 -----------------
 
+- Added `workflow.remove_state_transitions` function do remove transitions on a state and clean duplicates
+  [sgeulette]
 - Added more tests on cached methods.
   [sgeulette]